backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
python-multipart==0.0.22
pillow==11.0.0
piexif==0.0.14
